+++ ModFixes/Fixes.cs
@@ -28,6 +28,7 @@ namespace Polyfill.ModFixes
             new S1MapiPrefabLookup(),
             new S1MapiPrefabs(),
             new S1MapiInstancedTrees(),
+            new OverTheCounterNetworkLib(),
         };
 
         private static MelonPreferences_Entry<string> _disabled;